split the registration cypress e2e tests out into 3 steps and ensure that they run in a dedicated worker when run in github
change the number of times we intercept the recaptcha request in the registration_3.8.cy.ts test to 3 so that we can catch any request retries as well as just the first request
