+++ .github/workflows/cypress-ete.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        group: [1, 2, 3, 4, 5, 6, 7, 8, 9]
+        group: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
     timeout-minutes: 15
     services:
       redis:
